+++ app/routers/routers_posts.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, HTTPException, status
 import app.models.models as models
 from app.db.database import engine
 from app.db.configuration import db_dependency
-from app.models.models import PostBase
+from app.schemas.post_schema import PostBase
 
 router = APIRouter()
 models.Base.metadata.create_all(bind=engine)
