+++ app/schemas/post_schema.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class PostBase(BaseModel):
+    title: str
+    content: str
+    user_id: int
