+++ app/schemas/user_schema.py
@@ -1,0 +1,12 @@
+import re
+from pydantic import BaseModel, validator
+
+class UserBase(BaseModel):
+    id: int
+    username: str
+    password: str
+
+    #@validator('username')
+    #def validate_username(cls,value):
+    #    if not re.match(r'^[a-z0-9@]+$',value):
+    #        raise ValueError('Username format invalid')
