+++ requirements.txt
@@ -1,6 +1,8 @@
 annotated-types==0.7.0
 anyio==4.9.0
+cffi==1.17.1
 click==8.1.8
+cryptography==44.0.2
 ecdsa==0.19.1
 fastapi==0.115.12
 h11==0.14.0
@@ -10,6 +12,7 @@ packaging==24.2
 passlib==1.7.4
 pluggy==1.5.0
 pyasn1==0.4.8
+pycparser==2.22
 pydantic==2.11.3
 pydantic_core==2.33.1
 PyMySQL==1.1.1
